+++ pkg/cmd/pulumi/packagecmd/package_extract_schema.go
@@ -56,12 +56,8 @@ If a folder either the plugin binary must match the folder name (e.g. 'aws' and 
 			}()
 
 			parameters := &plugin.ParameterizeArgs{Args: args[1:]}
-			pkg, _, err := packages.SchemaFromSchemaSource(pctx, source, parameters,
+			spec, _, err := packages.SchemaFromSchemaSource(pctx, source, parameters,
 				cmdCmd.NewDefaultRegistry(cmd.Context(), pkgWorkspace.Instance, nil, cmdutil.Diag(), env.Global()))
-			if err != nil {
-				return err
-			}
-			spec, err := pkg.MarshalSpec()
 			if err != nil {
 				return err
 			}
@@ -77,6 +73,13 @@ If a folder either the plugin binary must match the folder name (e.g. 'aws' and 
 			if len(bytes) != n {
 				return fmt.Errorf("only wrote %d/%d bytes of the schema", len(bytes), n)
 			}
+
+			// Also try to bind the schema to warn about any diagnostics:
+			_, err = packages.BindSpec(*spec)
+			if err != nil {
+				return fmt.Errorf("failed to bind schema: %w", err)
+			}
+
 			return nil
 		},
 	}