+++ sdk/nodejs/tests/automation/util.ts
@@ -12,11 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { v4 as uuidv4 } from "uuid";
+import { randomUUID } from "crypto";
 
 /** @internal */
 export function getTestSuffix() {
-    return uuidv4();
+    return randomUUID();
 }
 
 /** @internal */